Safari-609.1.11.1/LayoutTests/fast/text/css-font-loading-arraybuffer-expected.txt

This test makes sure that an ArrayBufferView can be successfully passed to the FontFace constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fontFace1.status is "loaded"
PASS fontFace2.status is "loaded"
PASS document.getElementById('probe1').getBoundingClientRect().width is 100
PASS document.getElementById('probe2').getBoundingClientRect().width is 100
PASS successfullyParsed is true

TEST COMPLETE
l l
